requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-dysmsapi20170525==4.1.0
alibabacloud-tea==0.4.3
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.15
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-core-v3==2.13.33
aliyunsdkcore==1.0.3
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==0.10.0
multidict==6.4.3
numpy==2.0.2
openai==1.77.0
opencv-python==4.11.0.86
propcache==0.3.1
pycparser==2.22
pycrypto==2.6.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
